django-admin startproject arroyo
The Arroyo Bridge was dedicated in a ceremony officiated by Gov.
C. C. Young on Oct. 23, 1927, marking the completion of the first
structure on the Westwood campus of UCLA.

In the summer of 1947, the gully was filled to increase the amount
of useable property on the campus.

Today, the bridge’s arches remain hidden underground at Dickson
Plaza.

https://alumni.ucla.edu/ucla-history/ucla-history-42/

http://articles.latimes.com/2008/jul/13/local/me-then13

http://uclalibraryspecialcollections.tumblr.com/post/130819930661/then-the-arroyo-bridge-was-the-first-structure